fixes for tests

+++ tests/test_hypothesis.py
@@ -1,16 +1,14 @@
 from typing import List
 
 import hypothesis.strategies as st
-from hypothesis import given
-from hypothesis import settings
+from hypothesis import given, settings
 
-from pysmooth import Numeric
 from pysmooth.__main__ import smooth
 
 
 @settings(deadline=None)
 @given(x=st.lists(elements=st.integers(), min_size=3))
-def test_hypo_3RS3R_int(x: List[Numeric]) -> None:
+def test_hypo_3RS3R_int(x: List[float]) -> None:
     smooth(x=x, kind="3RS3R", twiceit=False, endrule="Tukey")
     smooth(x=x, kind="3RS3R", twiceit=False, endrule="copy")
 
@@ -20,7 +18,7 @@ def test_hypo_3RS3R_int(x: List[Numeric]) -> None:
 
 @settings(deadline=None)
 @given(x=st.lists(elements=st.integers(), min_size=3))
-def test_hypo_3RSS_int(x: List[Numeric]) -> None:
+def test_hypo_3RSS_int(x: List[float]) -> None:
     smooth(x=x, kind="3RSS", twiceit=False, endrule="Tukey")
     smooth(x=x, kind="3RSS", twiceit=False, endrule="copy")
 
@@ -30,7 +28,7 @@ def test_hypo_3RSS_int(x: List[Numeric]) -> None:
 
 @settings(deadline=None)
 @given(x=st.lists(elements=st.integers(), min_size=3))
-def test_hypo_3RSR_int(x: List[Numeric]) -> None:
+def test_hypo_3RSR_int(x: List[float]) -> None:
     smooth(x=x, kind="3RSR", twiceit=False, endrule="Tukey")
     smooth(x=x, kind="3RSR", twiceit=False, endrule="copy")
 
@@ -40,7 +38,7 @@ def test_hypo_3RSR_int(x: List[Numeric]) -> None:
 
 @settings(deadline=None)
 @given(x=st.lists(elements=st.integers(), min_size=3))
-def test_hypo_3R_int(x: List[Numeric]) -> None:
+def test_hypo_3R_int(x: List[float]) -> None:
     smooth(x=x, kind="3R", twiceit=False, endrule="Tukey")
     smooth(x=x, kind="3R", twiceit=False, endrule="copy")
 
@@ -50,7 +48,7 @@ def test_hypo_3R_int(x: List[Numeric]) -> None:
 
 @settings(deadline=None)
 @given(x=st.lists(elements=st.integers(), min_size=3))
-def test_hypo_3_int(x: List[Numeric]) -> None:
+def test_hypo_3_int(x: List[float]) -> None:
     smooth(x=x, kind="3", twiceit=False, endrule="Tukey")
     smooth(x=x, kind="3", twiceit=False, endrule="copy")
 
@@ -60,7 +58,7 @@ def test_hypo_3_int(x: List[Numeric]) -> None:
 
 @settings(deadline=None)
 @given(x=st.lists(elements=st.integers(), min_size=2, max_size=2))
-def test_hypo_3_2(x: List[Numeric]) -> None:
+def test_hypo_3_2(x: List[float]) -> None:
     smooth(x=x, kind="3", twiceit=False, endrule="Tukey")
     smooth(x=x, kind="3", twiceit=False, endrule="copy")
 
@@ -70,7 +68,7 @@ def test_hypo_3_2(x: List[Numeric]) -> None:
 
 @settings(deadline=None)
 @given(x=st.lists(elements=st.integers(), min_size=3))
-def test_hypo_S_int(x: List[Numeric]) -> None:
+def test_hypo_S_int(x: List[float]) -> None:
     smooth(x=x, kind="S", twiceit=False, endrule="Tukey")
     smooth(x=x, kind="S", twiceit=False, endrule="copy")
 
@@ -82,7 +80,7 @@ def test_hypo_S_int(x: List[Numeric]) -> None:
 @given(
     x=st.lists(elements=st.floats(allow_nan=False, allow_infinity=False), min_size=3)
 )
-def test_hypo_3RS3R_float(x: List[Numeric]) -> None:
+def test_hypo_3RS3R_float(x: List[float]) -> None:
 
     smooth(x=x, kind="3RS3R", twiceit=False, endrule="Tukey")
     smooth(x=x, kind="3RS3R", twiceit=False, endrule="copy")
@@ -95,7 +93,7 @@ def test_hypo_3RS3R_float(x: List[Numeric]) -> None:
 @given(
     x=st.lists(elements=st.floats(allow_nan=False, allow_infinity=False), min_size=3)
 )
-def test_hypo_3RSS_float(x: List[Numeric]) -> None:
+def test_hypo_3RSS_float(x: List[float]) -> None:
     smooth(x=x, kind="3RSS", twiceit=False, endrule="Tukey")
     smooth(x=x, kind="3RSS", twiceit=False, endrule="copy")
 
@@ -107,7 +105,7 @@ def test_hypo_3RSS_float(x: List[Numeric]) -> None:
 @given(
     x=st.lists(elements=st.floats(allow_nan=False, allow_infinity=False), min_size=3)
 )
-def test_hypo_3RSR_float(x: List[Numeric]) -> None:
+def test_hypo_3RSR_float(x: List[float]) -> None:
     smooth(x=x, kind="3RSR", twiceit=False, endrule="Tukey")
     smooth(x=x, kind="3RSR", twiceit=False, endrule="copy")
 
@@ -119,7 +117,7 @@ def test_hypo_3RSR_float(x: List[Numeric]) -> None:
 @given(
     x=st.lists(elements=st.floats(allow_nan=False, allow_infinity=False), min_size=3)
 )
-def test_hypo_3R_float(x: List[Numeric]) -> None:
+def test_hypo_3R_float(x: List[float]) -> None:
     smooth(x=x, kind="3R", twiceit=False, endrule="Tukey")
     smooth(x=x, kind="3R", twiceit=False, endrule="copy")
 
@@ -131,7 +129,7 @@ def test_hypo_3R_float(x: List[Numeric]) -> None:
 @given(
     x=st.lists(elements=st.floats(allow_nan=False, allow_infinity=False), min_size=3)
 )
-def test_hypo_3_float(x: List[Numeric]) -> None:
+def test_hypo_3_float(x: List[float]) -> None:
     smooth(x=x, kind="3", twiceit=False, endrule="Tukey")
     smooth(x=x, kind="3", twiceit=False, endrule="copy")
 
@@ -143,7 +141,7 @@ def test_hypo_3_float(x: List[Numeric]) -> None:
 @given(
     x=st.lists(elements=st.floats(allow_nan=False, allow_infinity=False), min_size=3)
 )
-def test_hypo_S_float(x: List[Numeric]) -> None:
+def test_hypo_S_float(x: List[float]) -> None:
     smooth(x=x, kind="S", twiceit=False, endrule="Tukey")
     smooth(x=x, kind="S", twiceit=False, endrule="copy")
